chore: divided base script into modules
