+++ Assets/PlayroomKit/PlayroomKit.cs
@@ -10,10 +10,9 @@ using System.Collections;
 
 namespace Playroom
 {
-    public class PlayroomKit
+    public partial class PlayroomKit
     {
-        private static bool isPlayRoomInitialized;
-
+        public static bool isPlayRoomInitialized;
 
         /// <summary>
         /// Required Mock Mode:
@@ -29,8 +28,6 @@ namespace Playroom
         private static Dictionary<string, object> MockDictionary = new();
 
         private static readonly Dictionary<string, Player> Players = new();
-
-
 
         [Serializable]
         public class InitOptions
@@ -829,7 +826,7 @@ namespace Playroom
             SetStateDictionary(key, jsonString, reliable);
         }
 
-        private static Dictionary<string, T> ParseJsonToDictionary<T>(string jsonString)
+        protected static Dictionary<string, T> ParseJsonToDictionary<T>(string jsonString)
         {
             var dictionary = new Dictionary<string, T>();
             var jsonNode = JSON.Parse(jsonString);
@@ -918,7 +915,7 @@ namespace Playroom
 #endif
         }
 
-        private static void MockSetState(string key, object value)
+        protected static void MockSetState(string key, object value)
         {
             if (MockDictionary.ContainsKey(key))
                 MockDictionary[key] = value;
@@ -926,7 +923,7 @@ namespace Playroom
                 MockDictionary.Add(key, value);
         }
 
-        private static T MockGetState<T>(string key)
+        protected static T MockGetState<T>(string key)
         {
             if (MockDictionary.TryGetValue(key, out var value) && value is T typedValue)
             {
@@ -1056,213 +1053,6 @@ namespace Playroom
         }
 
 
-        // RPC:
-        public enum RpcMode
-        {
-            ALL,
-            OTHERS,
-            HOST
-        }
-
-        private static Dictionary<string, Action<string, string>> rpcRegisterCallbacks = new();
-        private static List<string> rpcCalledEvents = new();
-
-        [DllImport("__Internal")]
-        private extern static void RpcRegisterInternal(string name, Action<string, string> rpcRegisterCallback, string onResponseReturn = null);
-
-        public static void RpcRegister(string name, Action<string, string> rpcRegisterCallback, string onResponseReturn = null)
-        {
-            if (IsRunningInBrowser())
-            {
-                rpcRegisterCallbacks.Add(name, rpcRegisterCallback);
-                RpcRegisterInternal(name, InvokeRpcRegisterCallBack, onResponseReturn);
-            }
-            else
-            {
-                Debug.LogError("[Mock Mode] RPC is currently not supported in Mock Mode!\n Please build the project to test RPC.");
-            }
-        }
-
-        [MonoPInvokeCallback(typeof(Action<string, string>))]
-        private static void InvokeRpcRegisterCallBack(string dataJson, string senderJson)
-        {
-            try
-            {
-                if (!Players.ContainsKey(senderJson))
-                {
-                    var player = new Player(senderJson);
-                    Players.Add(senderJson, player);
-                }
-
-            }
-            catch (Exception ex)
-            {
-                Debug.LogError(ex.Message);
-            }
-
-
-            List<string> updatedRpcCalledEvents = new();
-            // This state is required to update the called rpc events list, (Temp fix see RpcCall for more) 
-            string nameJson = GetState<string>("rpcCalledEventName");
-
-            JSONArray jsonArray = JSON.Parse(nameJson).AsArray;
-            foreach (JSONNode node in jsonArray)
-            {
-                string item = node.Value;
-                updatedRpcCalledEvents.Add(item);
-            }
-
-            foreach (string name in updatedRpcCalledEvents)
-            {
-                if (rpcRegisterCallbacks.TryGetValue(name, out Action<string, string> callback))
-                {
-                    callback?.Invoke(dataJson, senderJson);
-                }
-            }
-
-        }
-
-        [DllImport("__Internal")]
-        private extern static void RpcCallInternal(string name, string data, RpcMode mode, Action callbackOnResponse);
-
-        private static Dictionary<string, List<Action>> OnResponseCallbacks = new Dictionary<string, List<Action>>();
-
-        public static void RpcCall(string name, object data, RpcMode mode, Action callbackOnResponse)
-        {
-
-            if (IsRunningInBrowser())
-            {
-                string jsonData = ConvertToJson(data);
-                if (OnResponseCallbacks.ContainsKey(name))
-                {
-                    OnResponseCallbacks[name].Add(callbackOnResponse);
-                }
-                else
-                {
-                    OnResponseCallbacks.Add(name, new List<Action> { callbackOnResponse });
-                    if (!rpcCalledEvents.Contains(name))
-                    {
-                        rpcCalledEvents.Add(name);
-                    }
-                }
-                JSONArray jsonArray = new JSONArray();
-                foreach (string item in rpcCalledEvents)
-                {
-                    jsonArray.Add(item);
-                }
-                string jsonString = jsonArray.ToString();
-                /* 
-                This is requrired to sync the rpc events between all players, without this players won't know which event has been called.
-                this is a temporary fix, RPC's need to be handled within JS for better control.
-                */
-                SetState("rpcCalledEventName", jsonString, reliable: true);
-
-                RpcCallInternal(name, jsonData, mode, InvokeOnResponseCallback);
-            }
-            else
-            {
-                Debug.LogError("[Mock Mode] RPC Calls are not supported in Mock Mode! yet.\nPlease make a build to test RPC.");
-            }
-
-        }
-
-        // Default Mode
-        public static void RpcCall(string name, object data, Action callbackOnResponse)
-        {
-            RpcCall(name, data, RpcMode.ALL, callbackOnResponse);
-        }
-
-        [MonoPInvokeCallback(typeof(Action))]
-        private static void InvokeOnResponseCallback()
-        {
-            var namesToRemove = new List<string>();
-
-            foreach (string name in rpcCalledEvents)
-            {
-                try
-                {
-                    if (OnResponseCallbacks.TryGetValue(name, out List<Action> callbacks))
-                    {
-                        foreach (var callback in callbacks)
-                        {
-                            callback?.Invoke();
-                        }
-
-                        namesToRemove.Add(name);
-                    }
-                }
-                catch (Exception ex)
-                {
-                    Debug.LogError($"C#: Error in Invoking callback for RPC event name: '{name}': {ex.Message}");
-                }
-            }
-
-            foreach (var name in namesToRemove)
-            {
-                rpcCalledEvents.Remove(name);
-                OnResponseCallbacks.Remove(name);
-            }
-        }
-
-
-        private static string ConvertToJson(object data)
-        {
-            if (data == null)
-            {
-                return null;
-            }
-            else if (data.GetType().IsPrimitive || data is string)
-            {
-                return data.ToString();
-            }
-            if (data is Vector2 vector2)
-            {
-                return JsonUtility.ToJson(vector2);
-            }
-            else if (data is Vector3 vector3)
-            {
-                return JsonUtility.ToJson(vector3);
-            }
-            else if (data is Vector4 vector4)
-            {
-                return JsonUtility.ToJson(vector4);
-            }
-            else if (data is Quaternion quaternion)
-            {
-                return JsonUtility.ToJson(quaternion);
-            }
-            else
-            {
-                return ConvertComplexToJson(data);
-            }
-        }
-
-        private static string ConvertComplexToJson(object data)
-        {
-            if (data is IDictionary dictionary)
-            {
-                JSONObject dictNode = new JSONObject();
-                foreach (DictionaryEntry entry in dictionary)
-                {
-                    dictNode[entry.Key.ToString()] = ConvertToJson(entry.Value);
-                }
-                return dictNode.ToString();
-            }
-            else if (data is IEnumerable enumerable)
-            {
-                JSONArray arrayNode = new JSONArray();
-                foreach (object element in enumerable)
-                {
-                    arrayNode.Add(ConvertToJson(element));
-                }
-                return arrayNode.ToString();
-            }
-            else
-            {
-                Debug.Log($"{data} is '{data.GetType()}' which is currently not supported by RPC!");
-                return JSON.Parse("{}").ToString();
-            }
-        }
 
         [DllImport("__Internal")]
         private static extern void StartMatchmakingInternal(Action callback);
@@ -1287,628 +1077,7 @@ namespace Playroom
             startMatchmakingCallback?.Invoke();
         }
 
-        // Player class
 
-        public interface IPlayerInteraction
-        {
-            void InvokeOnQuitWrapperCallback();
-        }
-
-        public class Player : IPlayerInteraction
-        {
-
-            [Serializable]
-            public class Profile
-            {
-                [NonSerialized]
-                public UnityEngine.Color color;
-
-                public JsonColor jsonColor;
-                public string name;
-                public string photo;
-
-                [Serializable]
-                public class JsonColor
-                {
-                    public int r;
-                    public int g;
-                    public int b;
-                    public string hexString;
-                    public int hex;
-                }
-
-
-
-            }
-
-
-            public string id;
-            private static int totalObjects = 0;
-
-
-            public Player(string id)
-            {
-                this.id = id;
-                totalObjects++;
-
-                if (IsRunningInBrowser())
-                {
-                    // OnQuitCallbacks.Add(OnQuitDefaultCallback);
-                }
-                else
-                {
-                    if (!isPlayRoomInitialized)///
-                        Debug.LogError("[Mock Mode] Playroom not initialized yet! Please call InsertCoin.");
-                    else
-                        Debug.Log("Mock Player Created");
-                }
-            }
-
-            private List<Action<string>> OnQuitCallbacks = new();
-
-
-            private void OnQuitDefaultCallback()
-            {
-                if (!isPlayRoomInitialized)
-                {
-                    Debug.LogError("Playroom not initialized yet! Please call InsertCoin.");
-                }
-
-                Players.Remove(id);
-            }
-
-            [MonoPInvokeCallback(typeof(Action))]
-            private void OnQuitWrapperCallback()
-            {
-                if (OnQuitCallbacks != null)
-                    foreach (var callback in OnQuitCallbacks)
-                        callback?.Invoke(id);
-            }
-
-            void IPlayerInteraction.InvokeOnQuitWrapperCallback()
-            {
-                OnQuitWrapperCallback();
-            }
-
-            public Action OnQuit(Action<string> callback)
-            {
-                if (!isPlayRoomInitialized)
-                {
-                    Debug.LogError("PlayroomKit is not loaded!. Please make sure to call InsertCoin first.");
-                    return null;
-                }
-                else
-                {
-                    OnQuitCallbacks.Add(callback);
-
-                    void Unsubscribe()
-                    {
-                        OnQuitCallbacks.Remove(callback);
-                    }
-
-                    return Unsubscribe;
-                }
-            }
-
-            public void SetState(string key, int value, bool reliable = false)
-            {
-                if (IsRunningInBrowser())
-                {
-                    SetPlayerStateByPlayerId(id, key, value, reliable);
-                }
-                else
-                {
-                    if (!isPlayRoomInitialized)
-                    {
-                        Debug.LogError("[Mock Mode] Playroom not initialized yet! Please call InsertCoin.");
-                    }
-                    else
-                    {
-                        Debug.Log($"PlayerState Set! Key: {key}, Value: {value}");
-                        MockSetState(key, value);
-                    }
-                }
-            }
-
-
-            public void SetState(string key, float value, bool reliable = false)
-            {
-                if (IsRunningInBrowser())
-                {
-                    SetPlayerStateFloatByPlayerId(id, key, value.ToString(CultureInfo.InvariantCulture), reliable);
-                }
-                else
-                {
-                    if (!isPlayRoomInitialized)
-                    {
-                        Debug.LogError("[Mock Mode] Playroom not initialized yet! Please call InsertCoin.");
-                    }
-                    else
-                    {
-                        Debug.Log($"PlayerState Set! Key: {key}, Value: {value}");
-                        MockSetState(key, value);
-                    }
-                }
-            }
-
-            public void SetState(string key, bool value, bool reliable = false)
-            {
-                if (IsRunningInBrowser())
-                {
-                    SetPlayerStateByPlayerId(id, key, value, reliable);
-                }
-                else
-                {
-                    if (!isPlayRoomInitialized)
-                    {
-                        Debug.LogError("[Mock Mode] Playroom not initialized yet! Please call InsertCoin.");
-                    }
-                    else
-                    {
-                        Debug.Log($"PlayerState Set! Key: {key}, Value: {value}");
-                        MockSetState(key, value);
-                    }
-                }
-            }
-
-            public void SetState(string key, string value, bool reliable = false)
-            {
-                if (IsRunningInBrowser())
-                {
-                    SetPlayerStateStringById(id, key, value, reliable);
-                }
-                else
-                {
-                    if (!isPlayRoomInitialized)
-                    {
-                        Debug.LogError("[Mock Mode] Playroom not initialized yet! Please call InsertCoin.");
-                    }
-                    else
-                    {
-                        Debug.Log($"PlayerState Set! Key: {key}, Value: {value}");
-                        MockSetState(key, value);
-                    }
-                }
-            }
-
-            public void SetState(string key, object value, bool reliable = false)
-            {
-                if (IsRunningInBrowser())
-                {
-                    string jsonString = JsonUtility.ToJson(value);
-                    SetPlayerStateStringById(id, key, jsonString, reliable);
-                }
-                else
-                {
-                    if (!isPlayRoomInitialized)
-                    {
-                        Debug.LogError("[Mock Mode] Playroom not initialized yet! Please call InsertCoin.");
-                    }
-                    else
-                    {
-                        Debug.Log($"State Set! Key: {key}, Value: {value}");
-                        MockSetState(key, value);
-                    }
-                }
-            }
-
-            public T GetState<T>(string key)
-            {
-                if (IsRunningInBrowser())
-                {
-                    Type type = typeof(T);
-                    if (type == typeof(int)) return (T)(object)GetPlayerStateIntById(id, key);
-                    else if (type == typeof(float)) return (T)(object)GetPlayerStateFloatById(id, key);
-                    else if (type == typeof(bool)) return (T)(object)GetPlayerStateBoolById(id, key);
-                    else if (type == typeof(string)) return (T)(object)GetPlayerStateStringById(id, key);
-                    else if (type == typeof(Vector3)) return (T)(object)JsonUtility.FromJson<Vector3>(GetPlayerStateStringById(id, key));
-                    else if (type == typeof(Vector2)) return (T)(object)JsonUtility.FromJson<Vector2>(GetPlayerStateStringById(id, key));
-                    else if (type == typeof(Quaternion)) return (T)(object)JsonUtility.FromJson<Quaternion>(GetPlayerStateStringById(id, key));
-                    else throw new NotSupportedException($"Type {typeof(T)} is not supported by GetState");
-                }
-                else
-                {
-                    if (!isPlayRoomInitialized)
-                    {
-                        Debug.LogError("[Mock Mode] Playroom not initialized yet! Please call InsertCoin.");
-                        return default;
-                    }
-                    else
-                    {
-                        return MockGetState<T>(key);
-                    }
-                }
-            }
-
-
-            public Dictionary<string, T> GetState<T>(string key, bool isReturnDictionary = false)
-            {
-
-                if (IsRunningInBrowser() && isReturnDictionary)
-                {
-                    var jsonString = GetPlayerStateDictionary(id, key);
-                    return ParseJsonToDictionary<T>(jsonString);
-                }
-                else
-                {
-                    if (isPlayRoomInitialized)
-                    {
-                        if (isReturnDictionary)
-                        {
-                            return MockGetState<Dictionary<string, T>>(key);
-                        }
-                        else
-                        {
-                            return default;
-                        }
-                    }
-                    else
-                    {
-                        Debug.LogError("[Mock Mode] Playroom not initialized yet! Please call InsertCoin.");
-                        return default;
-                    }
-                }
-            }
-
-            private int GetPlayerStateInt(string key)
-            {
-                if (IsRunningInBrowser())
-                {
-                    return GetPlayerStateIntById(id, key);
-                }
-                else
-                {
-                    if (!isPlayRoomInitialized)
-                    {
-                        Debug.LogError("[Mock Mode] Playroom not initialized yet! Please call InsertCoin.");
-                        return default;
-                    }
-                    else
-                    {
-                        return MockGetState<int>(key);
-                    }
-                }
-            }
-
-            private float GetPlayerStateFloat(string key)
-            {
-                if (IsRunningInBrowser())
-                {
-                    return GetPlayerStateFloatById(id, key);
-                }
-                else
-                {
-                    if (!isPlayRoomInitialized)
-                    {
-                        Debug.LogError("[Mock Mode] Playroom not initialized yet! Please call InsertCoin.");
-                        return default;
-                    }
-                    else
-                    {
-                        return MockGetState<float>(key);
-                    }
-                }
-            }
-
-            private string GetPlayerStateString(string key)
-            {
-                if (IsRunningInBrowser())
-                {
-                    return GetPlayerStateStringById(id, key);
-                }
-                else
-                {
-                    if (!isPlayRoomInitialized)
-                    {
-                        Debug.LogError("[Mock Mode] Playroom not initialized yet! Please call InsertCoin.");
-                        return default;
-                    }
-                    else
-                    {
-                        return MockGetState<string>(key);
-                    }
-                }
-            }
-
-            private bool GetPlayerStateBool(string key)
-            {
-                if (IsRunningInBrowser())
-                {
-                    if (GetPlayerStateIntById(id, key) == 1)
-                    {
-                        return true;
-                    }
-                    else if (GetPlayerStateIntById(id, key) == 0)
-                    {
-                        return false;
-                    }
-                    else
-                    {
-                        Debug.LogError("GetPlayerStateByPlayerId: " + key + " is not a bool");
-                        return false;
-                    }
-                }
-                else
-                {
-                    if (!isPlayRoomInitialized)
-                    {
-                        Debug.LogError("[Mock Mode] Playroom not initialized yet! Please call InsertCoin.");
-                        return default;
-                    }
-                    else
-                    {
-                        return MockGetState<bool>(key);
-                    }
-                }
-            }
-
-            // Dictionaries:
-            public void SetState(string key, Dictionary<string, int> values, bool reliable = false)
-            {
-                if (IsRunningInBrowser())
-                {
-                    SetStateHelper(id, key, values, reliable);
-                }
-                else
-                {
-                    if (!isPlayRoomInitialized)
-                    {
-                        Debug.LogError("[Mock Mode] Playroom not initialized yet! Please call InsertCoin.");
-                    }
-                    else
-                    {
-                        Debug.Log($"PlayerState Set! Key: {key}, Value: {values}");
-                        MockSetState(key, values);
-                    }
-                }
-            }
-
-            public void SetState(string key, Dictionary<string, float> values, bool reliable = false)
-            {
-                if (IsRunningInBrowser())
-                {
-                    SetStateHelper(id, key, values, reliable);
-                }
-                else
-                {
-                    if (!isPlayRoomInitialized)
-                    {
-                        Debug.LogError("[Mock Mode] Playroom not initialized yet! Please call InsertCoin.");
-                    }
-                    else
-                    {
-                        Debug.Log($"PlayerState Set! Key: {key}, Value: {values}");
-                        MockSetState(key, values);
-                    }
-                }
-            }
-
-            public void SetState(string key, Dictionary<string, bool> values, bool reliable = false)
-            {
-                if (IsRunningInBrowser())
-                {
-                    SetStateHelper(id, key, values, reliable);
-                }
-                else
-                {
-                    if (!isPlayRoomInitialized)
-                    {
-                        Debug.LogError("[Mock Mode] Playroom not initialized yet! Please call InsertCoin.");
-                    }
-                    else
-                    {
-                        Debug.Log($"PlayerState Set! Key: {key}, Value: {values}");
-                        MockSetState(key, values);
-                    }
-                }
-            }
-
-            public void SetState(string key, Dictionary<string, string> values, bool reliable = false)
-            {
-                if (IsRunningInBrowser())
-                {
-                    SetStateHelper(id, key, values, reliable);
-                }
-                else
-                {
-                    if (!isPlayRoomInitialized)
-                    {
-                        Debug.LogError("[Mock Mode] Playroom not initialized yet! Please call InsertCoin.");
-                    }
-                    else
-                    {
-                        Debug.Log($"PlayerState Set! Key: {key}, Value: {values}");
-                        MockSetState(key, values);
-                    }
-                }
-            }
-
-            public void WaitForState(string StateKey, Action onStateSetCallback = null)
-            {
-                if (IsRunningInBrowser())
-                {
-                    WaitForPlayerStateInternal(id, StateKey, onStateSetCallback);
-                }
-            }
-
-            public void Kick(Action OnKickCallBack = null)
-            {
-                if (IsRunningInBrowser())
-                {
-                    OnKickCallBack = onKickCallBack;
-                    KickInternal(id, InvokeKickCallBack);
-                }
-                else
-                {
-                    if (!isPlayRoomInitialized)
-                    {
-                        Debug.LogError("[Mock Mode] Playroom not initialized yet! Please call InsertCoin.");
-                        return;
-                    }
-                    else
-                    {
-                        var player = GetPlayer(PlayerId);
-                        Players.Remove(player.id);
-                        onKickCallBack?.Invoke();
-                    }
-                }
-            }
-
-            [DllImport("__Internal")]
-            private static extern void SetPlayerStateByPlayerId(string playerID, string key, int value,
-                bool reliable = false);
-
-            [DllImport("__Internal")]
-            private static extern void SetPlayerStateFloatByPlayerId(string playerID, string key, string value,
-                bool reliable = false);
-
-            [DllImport("__Internal")]
-            private static extern void SetPlayerStateByPlayerId(string playerID, string key, bool value,
-                bool reliable = false);
-
-            [DllImport("__Internal")]
-            private static extern void SetPlayerStateStringById(string playerID, string key, string value,
-                bool reliable = false);
-
-            [DllImport("__Internal")]
-            private static extern string GetProfileByPlayerId(string playerID);
-
-            private static Profile ParseProfile(string json)
-            {
-                var jsonNode = JSON.Parse(json);
-                var profileData = new Profile();
-                profileData.jsonColor = new Profile.JsonColor
-                {
-                    r = jsonNode["color"]["r"].AsInt,
-                    g = jsonNode["color"]["g"].AsInt,
-                    b = jsonNode["color"]["b"].AsInt,
-                    hexString = jsonNode["color"]["hexString"].Value,
-                    hex = jsonNode["color"]["hex"].AsInt
-                };
-
-                ColorUtility.TryParseHtmlString(profileData.jsonColor.hexString, out UnityEngine.Color color1);
-                profileData.color = color1;
-                profileData.name = jsonNode["name"].Value;
-                profileData.photo = jsonNode["photo"].Value;
-
-                return profileData;
-            }
-
-            public Profile GetProfile()
-            {
-                if (IsRunningInBrowser())
-                {
-                    var jsonString = GetProfileByPlayerId(id);
-                    var profileData = ParseProfile(jsonString);
-                    return profileData;
-                }
-                else
-                {
-                    if (!isPlayRoomInitialized)
-                    {
-                        Debug.LogError("[Mock Mode] Playroom not initialized yet! Please call InsertCoin.");
-                        return default;
-                    }
-                    else
-                    {
-                        Profile.JsonColor mockJsonColor = new()
-                        {
-                            r = 166,
-                            g = 0,
-                            b = 142,
-                            hexString = "#a6008e"
-                        };
-                        ColorUtility.TryParseHtmlString(mockJsonColor.hexString, out UnityEngine.Color color1);
-                        var testProfile = new Profile()
-                        {
-                            color = color1,
-                            name = "CoolPlayTest",
-                            jsonColor = mockJsonColor,
-                            photo = "testPhoto"
-
-                        };
-                        return testProfile;
-                    }
-                }
-            }
-
-
-            private static Action onKickCallBack = null;
-
-            [MonoPInvokeCallback(typeof(Action))]
-            private static void InvokeKickCallBack()
-            {
-                onKickCallBack?.Invoke();
-            }
-
-            [DllImport("__Internal")]
-            private static extern void KickInternal(string playerID, Action onKickCallBack = null);
-
-            [DllImport("__Internal")]
-            private static extern void WaitForPlayerStateInternal(string playerID, string stateKey, Action onStateSetCallback = null);
-
-
-            private static bool GetPlayerStateBoolById(string id, string key)
-            {
-                if (IsRunningInBrowser())
-                {
-                    var stateValue = GetPlayerStateIntById(id, key);
-                    return stateValue == 1 ? true :
-                        stateValue == 0 ? false :
-                        throw new InvalidOperationException($"GetStateBool: {key} is not a bool");
-                }
-                else
-                {
-                    if (!isPlayRoomInitialized)
-                    {
-                        Debug.LogError("[Mock Mode] Playroom not initialized yet! Please call InsertCoin.");
-                        return false;
-                    }
-                    else
-                    {
-                        return MockGetState<bool>(key);
-                    }
-                }
-            }
-
-            [DllImport("__Internal")]
-            private static extern int GetPlayerStateIntById(string playerID, string key);
-
-            [DllImport("__Internal")]
-            private static extern float GetPlayerStateFloatById(string playerID, string key);
-
-            [DllImport("__Internal")]
-            private static extern string GetPlayerStateStringById(string playerID, string key);
-
-            // Helpers
-            [DllImport("__Internal")]
-            private static extern void SetPlayerStateDictionary(string playerID, string key, string jsonValues,
-                bool reliable = false);
-
-            [DllImport("__Internal")]
-            private static extern string GetPlayerStateDictionary(string playerID, string key);
-
-            private void SetStateHelper<T>(string id, string key, Dictionary<string, T> values, bool reliable = false)
-            {
-                var jsonObject = new JSONObject();
-
-                // Add key-value pairs to the JSON object
-                foreach (var kvp in values)
-                {
-                    // Convert the value to double before adding to JSONNode
-                    var value = Convert.ToDouble(kvp.Value);
-                    jsonObject.Add(kvp.Key, value);
-                }
-
-                // Serialize the JSON object to a string
-                var jsonString = jsonObject.ToString();
-
-                // Output the JSON string
-                SetPlayerStateDictionary(id, key, jsonString, reliable);
-            }
-
-
-        }
     }
