+++ Assets/PlayroomKit/modules.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 79a6747e9bd7d634ab4eda87b3573984
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
