+++ Assets/PlayroomKit/modules/Player.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dee2bc384854b2843a7a68de0436f253
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
