+++ Assets/PlayroomKit/modules/RPC.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: daadfc50e11096c4598b324f48edf1e9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
